Predict the values of the dependent variables using the quadVAR model

Predict the values of the dependent variables using the quadVAR model

Usage

# S3 method for class 'quadVAR'
predict(
  object,
  newdata = NULL,
  donotpredict = NULL,
  lowerbound = -Inf,
  upperbound = Inf,
  with_const = FALSE,
  ...
)

Arguments

object

A quadVAR object.

newdata

A data frame or tibble containing at least the values of the independent variables, dayvar, and beepvar (if used in model estimation). If NULL, the original data used to fit the model will be used.

donotpredict

NOT IMPLEMENTED YET! A character vector of the model names that are not used for prediction. Possible options include "AR", "VAR", "VAR_full", "quadVAR_full", "all_others", with NULL as the default. If set "all_others", then only a quadVAR model will be estimated. For datasets with large number of variables, you may set this parameter to "quadVAR_full" to save time.

lowerbound

A numeric value or a vector with the same length as the number of variables that specifies the lower bound of the predicted values. If the predicted value is less than this value, it will be replaced by this value. The default value is -Inf.

upperbound

A numeric value or a vector with the same length as the number of variables that specifies the upper bound of the predicted values. If the predicted value is greater than this value, it will be replaced by this value. The default value is Inf.

with_const

A logical value indicating whether to include the constant variables in the prediction. Those variables were automatically excluded in the estimation procedure. The default value is FALSE. When set to TRUE, the lowerbound and upperbound should be a vector with the same length as the number of variables in the model, including the constant variables. The values of the constant variables will be ignored though because their predicted values are always the same, which is the constant value in the input data.

...

Other arguments passed to the RAMP::predict.RAMP() function.

Value

A data frame or tibble containing the predicted values of the dependent variables. If a value cannot be predicted (e.g., because the corresponding previous time point is not in the data), it will be NA.